Move MACOS gui BulletClient.
