+++ smarts/core/utils/pybullet.py
@@ -17,8 +17,8 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-import functools
-import inspect
+from multiprocessing import Pipe, Process
+from multiprocessing.connection import Connection
 
 from smarts.core.utils.core_logging import suppress_output
 
@@ -61,3 +61,52 @@ class SafeBulletClient(bullet_client.BulletClient):
         if name in {"__deepcopy__", "__getstate__", "__setstate__"}:
             raise RuntimeError(f"{self.__class__} does not allow `{name}`")
         return super().__getattr__(name)
+
+
+class BulletClientMACOS:
+    """This wrapper class is a hack for `macOS` where running PyBullet in GUI mode,
+    alongside Panda3D segfaults. It seems due to running two OpenGL applications
+    in the same process. Here we spawn a process to run PyBullet and forward all
+    calls to it over unix pipes.
+
+    N.B. This class can be directly subbed-in for pybullet_utils's BulletClient
+    but your application must start from a,
+
+        if __name__ == "__main__:":
+            # https://turtlemonvh.github.io/python-multiprocessing-and-corefoundation-libraries.html
+            import multiprocessing as mp
+            mp.set_start_method('spawn', force=True)
+    """
+
+    def __init__(self, bullet_connect_mode=GUI):
+        self._parent_conn, self._child_conn = Pipe()
+        self.process = Process(
+            target=BulletClientMACOS.consume,
+            args=(
+                bullet_connect_mode,
+                self._child_conn,
+            ),
+        )
+        self.process.start()
+
+    def __getattr__(self, name):
+        def wrapper(*args, **kwargs):
+            self._parent_conn.send((name, args, kwargs))
+            return self._parent_conn.recv()
+
+        return wrapper
+
+    @staticmethod
+    def consume(bullet_connect_mode, connection: Connection):
+        """Builds a child pybullet process.
+        Args:
+            bullet_connect_mode: The type of bullet process.
+            connection: The child end of a pipe.
+        """
+        # runs in sep. process
+        client = SafeBulletClient(bullet_connect_mode)
+
+        while True:
+            method, args, kwargs = connection.recv()
+            result = getattr(client, method)(*args, **kwargs)
+            connection.send(result)
